Update high_resolution_filament_sensor.py to new klipper v0.13 syntax
In one of the recent klipper updates, extruder.move was changed to extruder.process_move. This broke the code for the filament sensor and didn't allow the printer to turn on. Updating the places in the code with extruder.move to extruder.process_move fixed this issue. Further issues were caused by _capture_extruder_move now requiring 4 inputs rather than 3. I added "ea_index" to the inputs and the return for that section in the code so it matches that in the toolhead.py code in klipper.